refactor: move legend styles into the style folder
