+++ giraffe/src/style/legend.ts
